Moved code for settings to separate module
